@d-netto 2024-01-21
Note: this was tested on x86_64.

It seems like the M2, for instance, has a cache-line of 128-bytes, so we might need to increase the padding size on this platform.